+++ src/Fields/Section/Section.php
@@ -50,7 +50,7 @@ class Section extends TeaFields
      * @param array $content Contains all data
      * @param array $post Contains all post data
      *
-     * @since 1.5.2-11
+     * @since 1.5.2.11
      */
     public function templatePages($content, $post = array(), $prefix = '')
     {